+++ src/main/java/dev/project/scholar_ai/security/SecurityConfig.java
@@ -53,10 +53,8 @@ public class SecurityConfig {
                 .authorizeHttpRequests(authorizeRequests -> authorizeRequests
                         .requestMatchers(
                                 "/api/test/test",
-                                "/api/v1/auth/login",
-                                "/api/v1/auth/register",
-                                "/api/v1/auth/refresh",
-                                "/api/v1/auth/social-login",
+                                "/api/v1/auth/**",
+                                "/api/v1/auth/social/**",
                                 "/api/demo/**",
                                 "/actuator/**",
                                 "/docs",
